Major refactor#1
